+++ dup_fmt/__init__.py
@@ -4,6 +4,10 @@
 # license information.
 # --------------------------------------------------------------------------
 
+from .__about__ import (
+    __version__,
+    __version_tuple__,
+)
 from .formatter import (
     FORMATTERS,
     FORMATTERS_ADJUST,
@@ -19,4 +23,29 @@ from .formatter import (
     ReturnPrioritiesType,
     Serial,
     Version,
+    make_order_fmt,
+)
+from .objects import (
+    relativeserial,
+)
+
+__all__ = (
+    "relativeserial",
+    "FORMATTERS",
+    "FORMATTERS_ADJUST",
+    "Constant",
+    "ConstantType",
+    "Datetime",
+    "EnvConstant",
+    "Formatter",
+    "FormatterGroup",
+    "Naming",
+    "OrderFormatter",
+    "ReturnFormattersType",
+    "ReturnPrioritiesType",
+    "Serial",
+    "Version",
+    "make_order_fmt",
+    "__version__",
+    "__version_tuple__",
 )