+++ tests/test_formatter_order.py
@@ -31,7 +31,14 @@ class OrderFormatTestCase(unittest.TestCase):
         self.fmt_order4 = fmt.OrderFormatter({"version": self.vs2})
         self.fmt_order5 = fmt.OrderFormatter({"serial": self.sr_dict})
         self.fmt_order6 = fmt.OrderFormatter({"serial": self.sr_dict2})
-        # self.fmt_order7 = fmt.OrderFormatter({"number": self.sr_dict2})
+        self.fmt_order7 = fmt.OrderFormatter(
+            {"timestamp": self.dt},
+            auto_serial=True,
+        )
+        self.fmt_order9 = fmt.OrderFormatter(
+            {"timestamp": self.dt, "serial": self.sr_dict},
+            auto_serial=True,
+        )
 
     def test_order_raise_name_init(self):
         raise_respec: Dict[str, Dict[Any, Any]] = {"number": self.sr_dict2}
@@ -67,8 +74,18 @@ class OrderFormatTestCase(unittest.TestCase):
             "(timestamp=['2022-01-01 00:00:00.000'])", self.fmt_order.__str__()
         )
 
+        self.assertEqual(
+            "(timestamp=['2022-01-01 00:00:00.000'], serial=['0'])",
+            self.fmt_order7.__str__(),
+        )
+
+        self.assertEqual(
+            "(timestamp=['2022-01-01 00:00:00.000'], serial=['2'])",
+            self.fmt_order9.__str__(),
+        )
+
     def test_order_timestamp(self):
-        self.assertTrue(self.fmt_order == self.fmt_order)
+        self.assertFalse(self.fmt_order == self.fmt_order2)
         self.assertFalse(self.fmt_order > self.fmt_order2)
         self.assertTrue(self.fmt_order < self.fmt_order2)
 
@@ -97,7 +114,7 @@ class OrderFormatTestCase(unittest.TestCase):
             self.fmt_order3.adjust_timestamp(metrics={"months": 1})
         self.assertTrue(
             (
-                "with 'timestamp', order file object does not have "
+                "with 'timestamp', order formatter object does not have "
                 "`timestamp` in name formatter"
             )
             in str(context.exception)
@@ -119,7 +136,7 @@ class OrderFormatTestCase(unittest.TestCase):
             self.fmt_order.adjust_version("*.1.3")
         self.assertTrue(
             (
-                "with 'version', order file object does not have "
+                "with 'version', order formatter object does not have "
                 "`version` in name formatter"
             )
             in str(context.exception)
@@ -135,7 +152,7 @@ class OrderFormatTestCase(unittest.TestCase):
             self.fmt_order.adjust_serial(1)
         self.assertTrue(
             (
-                "with 'serial', order file object does not have "
+                "with 'serial', order formatter object does not have "
                 "`serial` in name formatter"
             )
             in str(context.exception)